+++ pcbnew/widgets/appearance_controls.cpp
@@ -302,7 +302,7 @@ void NET_GRID_TABLE::HideOtherNets( const NET_GRID_ENTRY& aNet )
 void NET_GRID_TABLE::updateNetVisibility( const NET_GRID_ENTRY& aNet )
 {
     const TOOL_ACTION& action = aNet.visible ? PCB_ACTIONS::showNet : PCB_ACTIONS::hideNet;
-    m_frame->GetToolManager()->RunAction( action, true, aNet.code );
+    m_frame->GetToolManager()->RunAction( action, true, static_cast<intptr_t>( aNet.code ) );
 }
 
 
@@ -2488,21 +2488,24 @@ void APPEARANCE_CONTROLS::onNetContextMenu( wxCommandEvent& aEvent )
 
     case ID_HIGHLIGHT_NET:
     {
-        m_frame->GetToolManager()->RunAction( PCB_ACTIONS::highlightNet, true, net.code );
+        m_frame->GetToolManager()->RunAction( PCB_ACTIONS::highlightNet, true,
+                                              static_cast<intptr_t>( net.code ) );
         m_frame->GetCanvas()->Refresh();
         break;
     }
 
     case ID_SELECT_NET:
     {
-        m_frame->GetToolManager()->RunAction( PCB_ACTIONS::selectNet, true, net.code );
+        m_frame->GetToolManager()->RunAction( PCB_ACTIONS::selectNet, true,
+                                              static_cast<intptr_t>( net.code ) );
         m_frame->GetCanvas()->Refresh();
         break;
     }
 
     case ID_DESELECT_NET:
     {
-        m_frame->GetToolManager()->RunAction( PCB_ACTIONS::deselectNet, true, net.code );
+        m_frame->GetToolManager()->RunAction( PCB_ACTIONS::deselectNet, true,
+                                              static_cast<intptr_t>( net.code ) );
         m_frame->GetCanvas()->Refresh();
         break;
     }
@@ -2553,7 +2556,7 @@ void APPEARANCE_CONTROLS::showNetclass( const wxString& aClassName, bool aShow )
         {
             if( net->GetNetClass() == defaultClass )
             {
-                manager->RunAction( action, true, net->GetNetCode() );
+                manager->RunAction( action, true, static_cast<intptr_t>( net->GetNetCode() ) );
 
                 int row = m_netsTable->GetRowByNetcode( net->GetNetCode() );
 
@@ -2571,7 +2574,7 @@ void APPEARANCE_CONTROLS::showNetclass( const wxString& aClassName, bool aShow )
             if( NETINFO_ITEM* net = nets.GetNetItem( member ) )
             {
                 int code = net->GetNetCode();
-                manager->RunAction( action, true, code );
+                manager->RunAction( action, true, static_cast<intptr_t>( code ) );
 
                 int row = m_netsTable->GetRowByNetcode( code );
 
@@ -2746,7 +2749,7 @@ void APPEARANCE_CONTROLS::onNetclassContextMenu( wxCommandEvent& aEvent )
                             if( !aItem )
                                 return;
 
-                            int code = aItem->GetNetCode();
+                            intptr_t code = static_cast<intptr_t>( aItem->GetNetCode() );
                             m_frame->GetToolManager()->RunAction( action, true, code );
                         } );
             }
